+++ src/app/job.service.ts
@@ -12,14 +12,14 @@ export class JobService {
 
   createJob(newJob: Job){
     ipcRenderer.send('async-job-create', newJob)
-    ipcRenderer.on('async-job-create-reply', (event, arg) => {
+    ipcRenderer.once('async-job-create-reply', (event, arg) => {
       console.log(arg)
     })
   }
 
   updateJob(modifiedJob: Job) {
     ipcRenderer.send('async-job-update', modifiedJob)
-    ipcRenderer.on('async-job-update-reply', (event, arg) => {
+    ipcRenderer.once('async-job-update-reply', (event, arg) => {
       console.log(arg)
     })
   }
@@ -29,16 +29,68 @@ export class JobService {
       id: jobId
     }
     ipcRenderer.send('async-job-remove', jobData)
-    ipcRenderer.on('async-job-remove-reply', (event, arg) => {
+    ipcRenderer.once('async-job-remove-reply', (event, arg) => {
       console.log(arg)
     })
   }
 
   getAllJobs() {
     ipcRenderer.send('async-job-get-all');
-    ipcRenderer.on('async-job-get-all-reply',(event, arg) =>{
+    ipcRenderer.once('async-job-get-all-reply',(event, arg) =>{
       this.jobList = [];
       console.log(arg)
+      arg.forEach(element => {
+        this.jobList.push(element);
+      });
+    })
+    return this.jobList;
+  }
+
+  getJobByEmployee(employeeId: string){
+    const employee = {
+      employeeId: employeeId
+    }
+    ipcRenderer.send('async-job-get-by-employee', employee);
+    ipcRenderer.once('async-job-get-by-employee-reply', (event, arg)=> {
+      this.jobList = [];
+      console.log(arg);
+      arg.forEach(element => {
+        this.jobList.push(element);
+      });
+    })
+    return this.jobList;
+  }
+
+  getAvailableJobs(){
+    ipcRenderer.send('async-job-get-available');
+    ipcRenderer.once('async-job-get-available-reply', (event, arg) => {
+      this.jobList = [];
+      arg.forEach(element => {
+        this.jobList.push(element);
+      });
+    })
+    return this.jobList;
+  }
+
+  getEmployeeJobHistory(employeeId: string){
+    const employee = {
+      employeeId: employeeId
+    }
+    ipcRenderer.send('async-job-get-employee-history', employee);
+    ipcRenderer.once('async-job-get-employee-history-reply', (event,arg)=> {
+      this.jobList = [];
+      console.log(arg);
+      arg.forEach(element => {
+        this.jobList.push(element);
+      });
+    })
+    return this.jobList;
+  }
+
+  getJobsForReview(){
+    ipcRenderer.send('async-job-get-review');
+    ipcRenderer.once('async-job-get-review-reply', (event,arg)=>{
+      this.jobList = [];
       arg.forEach(element => {
         this.jobList.push(element);
       });